Moved customFieldNameValidation.js into validations folder
