+++ src/Widgets/FormInputFieldWidget/FormInputFieldWidgetEditingConfig.js
@@ -1,6 +1,6 @@
 import * as Scrivito from "scrivito";
 import formInputFieldWidgetIcon from "../../assets/images/form_input_field_widget.svg";
-import { customFieldNameValidation } from "../FormContainerWidget/customFieldNameValidation";
+import { customFieldNameValidation } from "../FormContainerWidget/validations/customFieldNameValidation";
 import { typeValidation } from "../FormContainerWidget/typeValidation";
 
 Scrivito.provideEditingConfig("FormInputFieldWidget", {
